+++ src/main/java/org/enchere/bo/Retrait.java
@@ -68,11 +68,7 @@ public class Retrait {
 
     @Override
     public String toString() {
-        return "Retrait{" +
-                "rue='" + rue + '\'' +
-                ", code_postal=" + code_postal +
-                ", ville='" + ville + '\'' +
-                '}';
+        return  rue + ", " + code_postal + " " + ville;
     }
 
 }
